+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * Expose icons as React compontner
 * Add CLI tool to eject icons into a specified directory
 * Add Documentation site with examples for every icon
+* Expose SVG sprite of all icons